Create a no-op exposure runner
