+++ core/dbt/runner/exposure_runner.py
@@ -1,0 +1,7 @@
+from dbt.runner.no_op_runner import NoOpRunner
+
+
+class ExposureRunner(NoOpRunner):
+    @property
+    def description(self) -> str:
+        return f"exposure {self.node.name}"
