+++ core/dbt/runner/saved_query_runner.py
@@ -1,0 +1,7 @@
+from dbt.runner.no_op_runner import NoOpRunner
+
+
+class SavedQueryRunner(NoOpRunner):
+    @property
+    def description(self) -> str:
+        return f"saved query {self.node.name}"
