+++ tests/unit/task/test_build.py
@@ -1,5 +1,5 @@
 from dbt.contracts.graph.nodes import SavedQuery
-from dbt.task.build import SavedQueryRunner
+from dbt.runner import SavedQueryRunner
 
 
 def test_saved_query_runner_on_skip(saved_query: SavedQuery):
